+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 ##### Please refer to the [release notes](https://github.com/wso2/ballerina-plugin-vscode/blob/main/docs/release-notes/3.3.0-release-note.md) for further information.
 
+## **3.3.3**
+### Fixed
+- Diagrams not loading with VSCode v1.73 issue.
+
+## **3.3.2**
+### Improved
+- The low-code diagram editor
+
 ## **3.3.1**
 ### Fixed
 - Try it button not working for services with comments issue.
@@ -9,6 +17,7 @@
 ### Added
 - Visual Data Mapper - Helps you write and visualize data transformations easily
 - GraphQL Tryit - Facilitates  trying out the GraphQL services with the integrated client
+- Project Design View (Experimental) - Allows you to visualize service interactions in your project
 
 ### Improved
 - Improved record editor - Provides a better editing experience with suggestions